Extract `TestContainer` to `Elsa.Testing.Shared.Activities` for reuse across test projects.
