+++ src/common/Elsa.Testing.Shared/Activities/TestContainer.cs
@@ -1,9 +1,12 @@
-namespace Elsa.Workflows.IntegrationTests.Activities.Container;
+using Elsa.Workflows;
+using Elsa.Workflows.Activities;
+
+namespace Elsa.Testing.Shared.Activities;
 
 /// <summary>
 /// Concrete implementation of Container for testing that schedules children sequentially.
 /// </summary>
-public class TestContainer : Elsa.Workflows.Activities.Container
+public class TestContainer : Container
 {
     private const string CurrentIndexProperty = "CurrentIndex";
